# internal/api/openapi.yaml
basePath: /api
definitions:
...
  api.createUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
...
  api.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.updateUserRoleRequest:
    properties:
      role:
        type: string
    type: object
  api.userResponse:
    properties:
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: Self-hosted Railway-style deploy platform for Kubernetes. Manages projects,
    apps, deployments, env vars, secrets, domains, and git providers.
  title: Mortise API
  version: 0.1.0
paths:
...
  /admin/users:
    get:
      description: Returns all platform users. Admin-only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.userResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new platform user. Admin-only.
      parameters:
      - description: User details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a user
      tags:
      - users
  /admin/users/{email}:
    delete:
      description: Removes a platform user. Admin-only.
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Changes a platform user's role. Admin-only.
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.updateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a user's role
      tags:
      - users
...
  /projects/{project}/apps/{app}/traffic:
    get:
      description: Proxies to the configured traffic adapter to return time-series
        request rate and latency data
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      - description: App name
        in: path
        name: app
        required: true
        type: string
      - description: 'Environment name (default: production)'
        in: query
        name: env
        type: string
      - description: Start timestamp (unix seconds)
        in: query
        name: start
        required: true
        type: string
      - description: End timestamp (unix seconds)
        in: query
        name: end
        required: true
        type: string
      - description: 'Step interval in seconds (default: 5)'
        in: query
        name: step
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Traffic data or availability status
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Get traffic history for an app
      tags:
      - traffic
  /projects/{project}/apps/{app}/traffic/current:
    get:
      description: Returns the most recent traffic data bucket from the adapter
      parameters:
      - description: Project name
        in: path
        name: project
        required: true
        type: string
      - description: App name
        in: path
        name: app
        required: true
        type: string
      - description: 'Environment name (default: production)'
        in: query
        name: env
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Traffic data or availability status
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - BearerAuth: []
      summary: Get current traffic for an app
      tags:
      - traffic
...
securityDefinitions:
  BearerAuth:
    description: JWT bearer token from POST /api/auth/login
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"